Updated addressing various review comments from the PR Changed struct definitions to be consistent with code base Changed Arm to 32-bit Fixed typo's Remove SEM_MAX_VALUE reference Fixed return type mismatches # <Type>: <Subject text> # 50-char limit ######### # 72-character lines for description ###################################
# Why:

# Related Issues:

# References:

# Run git config commit.template .gitmessage
# Use git commit when committing files, and this template will be used.

# Example commit message:
# chore: Add commit message template

# Capitalize subject line.
# Write capitalized imperative statement with no period for subject
# List details of significant changes below the subject line
# Write details as capitalized imperative statements
# Provide the below sections for further context

# Why:

# A short imperative subject helps when viewing the log, especially when
# viewing one-line formatted logs.  Clear writing of what changed helps
# give a clean understanding of context when reading blame messages and
# when looking at a specific commit.

# Related Issues:
# Link to related issue numbers (hyper link if applicable), e.g.:
# git issue 123

# References:
# List relevant references e.g.:
# https://chris.beams.io/posts/git-commit/

# Type can be
#    chore    (updating grunt tasks etc; no production code change)
#    docs     (changes to documentation, comments.  No logic change)
#    feat     (New feature)
#    fix      (Bug fix, or other change related to tracked issue)
#    merge    (Merge of one branch into another, no other change)
#    refactor (refactoring production code)
#    revert   (Revert of a previous commit, no other change)
#    style    (formatting, missing semi colons, etc; no code change)
#    test     (adding or refactoring tests; no production code change)
#    tool     (adding or refactoring tools; no production code change)

+++ lib/utils/platform/iot_clock_afr.c
@@ -61,7 +61,7 @@
 
 /*-----------------------------------------------------------*/
 
-/*  Private Callback function for timer expirty, delegate work to a Task to free 
+/*  Private Callback function for timer expiry, delegate work to a Task to free 
     up the timer task for managing other timers */
 static void prvTimerCallback( TimerHandle_t xTimerHandle )
 {
@@ -87,7 +87,7 @@ bool IotClock_GetTimestring( char * const pBuffer,
                                 size_t * const pTimestringLength )
 {
     uint64_t milliSeconds = IotClock_GetTimeMs( );
-    size_t timestringLength = 0;
+    int timestringLength = 0;
 
     /* Convert the localTime struct to a string. */
     timestringLength = snprintf( pBuffer, bufferSize, "%llu",  milliSeconds );
@@ -126,7 +126,7 @@ uint64_t IotClock_GetTimeMs( void )
     ullTickCount += xCurrentTime.xTimeOnEntering;
 
     /* Return the ticks converted to Milliseconds */
-    return ( int64_t ) ullTickCount * _MILLISECONDS_PER_TICK;
+    return ullTickCount * _MILLISECONDS_PER_TICK;
 }
 
 /*-----------------------------------------------------------*/
@@ -189,8 +189,8 @@ void IotClock_TimerDestroy( IotTimer_t * const pTimer )
 /*-----------------------------------------------------------*/
 
 bool IotClock_TimerArm( IotTimer_t * const pTimer,
-                           uint64_t relativeTimeoutMs,
-                           uint64_t periodMs )
+                           uint32_t relativeTimeoutMs,
+                           uint32_t periodMs )
 {
     _timerInfo_t * pTimerInfo = ( _timerInfo_t * ) pTimer;
